configs: add CONFIG_SENSORS_IIO_HWMON=m
commit  237e899fed1ba39c068e2ff35f41c9b7c3b1fc41 from
https://github.com/raspberrypi/linux.git rpi-5.4.y

The hwmon module for the ads1015 was removed from the kernel
in favor of the IIO version.  This patch adds the bridge driver
so that userland consumers of the hwmon data can continue to
do so via that interface.

Signed-off-by: Annaliese McDermond <nh6z@nh6z.net>
Signed-off-by: Meng Li <Meng.Li@windriver.com>