dtoverlays: Add the iio_hwmon driver to correct ADC issues
commit  8ae5f3e1ecf0c41cb788d90c048a10c27e8aaa02 from
https://github.com/raspberrypi/linux.git rpi-5.4.y

The Linux kernel maintainers removed the hwmon driver for the
ads1015 used on this board.  They deprecated it in favor of using
the IIO version of the driver with the iio_hwmon bridge.  This
patch updates the DRAWS dtoverlay to support that usage.

Signed-off-by: Annaliese McDermond <nh6z@nh6z.net>
Signed-off-by: Meng Li <Meng.Li@windriver.com>