ARM: dts: bcm2708.dtsi: Don't delete the cpus node
commit  a5c01bebc340e223c0ce5328242c867fdbffd08a from
https://github.com/raspberrypi/linux.git rpi-5.4.y

The cpus node was originally deleted to match the then downstream
version of the BCM2835 DTS files, but doing so doesn't seem to make
any material difference. Its presence is necessary for the CPUFREQ_DT
support that is planned in the near future, so remove the deletion.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>