perf/x86/intel/uncore: Factor out uncore_pci_get_dev_die_info()
commit fe6507338d635f283e9618b5eaa35f503a8c375b upstream.

The socket and die information is required to register/unregister a PMU
in the uncore PCI sub driver. The codes, which get the socket and die
information from a BUS number, can be shared.

Factor out uncore_pci_get_dev_die_info(), which will be used later.

There is no functional change.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: https://lkml.kernel.org/r/1600094060-82746-2-git-send-email-kan.liang@linux.intel.com
Signed-off-by: Liwei Song <liwei.song@windriver.com>